Suppression erreur sur admin monstre

+++ web/www/jeu_test/admin_type_monstre_edit.php
@@ -17,7 +17,7 @@ $rep = opendir($baseimage);
 $images_list="";
 $img=0;
 while (false !== ($filename = readdir($rep))) {
-    $avatar_perso_cod = 1* substr($filename, 0, -4);
+    $avatar_perso_cod = 1* (int) substr($filename, 0, -4);
     if ($avatar_perso_cod==0) {   // si le pattern correspond, c'est une image de perso (pas de monstre) on ne l'affiche pas!
         $imagesize = @getimagesize($baseimage.'/'.$filename) ;
         if (($imagesize[0] > 28) && ($imagesize[1] > 28)) {     // on ne prend que des images de taille raisonnable
